:wrench: Make animated icons appear only in session details. I gave up applying animated icons in the list because I could not manage the state correctly.
