+++ feature/sessions/src/main/java/io/github/droidkaigi/confsched2023/sessions/component/TimetableItemDetailBottomAppBar.kt
@@ -1,9 +1,15 @@
 package io.github.droidkaigi.confsched2023.sessions.component
 
+import android.os.Build
+import androidx.compose.animation.graphics.ExperimentalAnimationGraphicsApi
+import androidx.compose.animation.graphics.res.animatedVectorResource
+import androidx.compose.animation.graphics.res.rememberAnimatedVectorPainter
+import androidx.compose.animation.graphics.vector.AnimatedImageVector
+import androidx.compose.foundation.clickable
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Bookmark
-import androidx.compose.material.icons.filled.BookmarkBorder
 import androidx.compose.material.icons.filled.Share
+import androidx.compose.material.icons.outlined.BookmarkBorder
 import androidx.compose.material3.BottomAppBar
 import androidx.compose.material3.BottomAppBarDefaults
 import androidx.compose.material3.FloatingActionButton
@@ -12,9 +18,14 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
 import io.github.droidkaigi.confsched2023.designsystem.preview.MultiLanguagePreviews
 import io.github.droidkaigi.confsched2023.designsystem.preview.MultiThemePreviews
 import io.github.droidkaigi.confsched2023.designsystem.theme.KaigiTheme
@@ -53,20 +64,24 @@ fun TimetableItemDetailBottomAppBar(
         },
         floatingActionButton = {
             FloatingActionButton(
-                onClick = { onBookmarkClick(timetableItem) },
+                onClick = {
+                    // NOOP ,
+                },
                 modifier = Modifier.testTag(TimetableItemDetailBookmarkIconTestTag),
                 containerColor = BottomAppBarDefaults.bottomAppBarFabColor,
                 elevation = FloatingActionButtonDefaults.bottomAppBarFabElevation(),
             ) {
-                if (isBookmarked) {
-                    Icon(
-                        imageVector = Icons.Filled.Bookmark,
-                        contentDescription = SessionsStrings.RemoveFromFavorites.asString(),
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
+                    AnimatedBookmarkIcon(
+                        isBookmarked = isBookmarked,
+                        timetableItem = timetableItem,
+                        onClick = onBookmarkClick,
                     )
                 } else {
-                    Icon(
-                        imageVector = Icons.Filled.BookmarkBorder,
-                        contentDescription = SessionsStrings.AddToFavorites.asString(),
+                    BookmarkIcon(
+                        isBookmarked = isBookmarked,
+                        timetableItem = timetableItem,
+                        onClick = onBookmarkClick,
                     )
                 }
             }
@@ -74,8 +89,63 @@ fun TimetableItemDetailBottomAppBar(
     )
 }
 
+@OptIn(ExperimentalAnimationGraphicsApi::class)
+@Composable
+fun AnimatedBookmarkIcon(
+    isBookmarked: Boolean,
+    timetableItem: TimetableItem,
+    onClick: (TimetableItem) -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    var atEnd by remember { mutableStateOf(false) }
+    val animatedBookmarkIcon = AnimatedImageVector.animatedVectorResource(
+        id = if (isBookmarked) {
+            R.drawable.animated_bookmark_icon_reverse
+        } else {
+            R.drawable.animated_bookmark_icon
+        }
+    )
+    Icon(
+        painter = rememberAnimatedVectorPainter(animatedBookmarkIcon, atEnd),
+        contentDescription = if (isBookmarked) {
+            SessionsStrings.RemoveFromFavorites.asString()
+        } else {
+            SessionsStrings.AddToFavorites.asString()
+        },
+        modifier = modifier
+            .clickable {
+                atEnd = atEnd.not()
+                onClick(timetableItem)
+            },
+    )
+}
+
+@Composable
+fun BookmarkIcon(
+    isBookmarked: Boolean,
+    timetableItem: TimetableItem,
+    onClick: (TimetableItem) -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    Icon(
+        imageVector = if (isBookmarked) {
+            Icons.Filled.Bookmark
+        } else {
+            Icons.Outlined.BookmarkBorder
+        },
+        contentDescription = if (isBookmarked) {
+            SessionsStrings.RemoveFromFavorites.asString()
+        } else {
+            SessionsStrings.AddToFavorites.asString()
+        },
+        modifier = modifier
+            .clickable { onClick(timetableItem) },
+    )
+}
+
 @MultiThemePreviews
 @MultiLanguagePreviews
+@Preview(apiLevel = 30)
 @Composable
 fun TimetableItemDetailBottomAppBarPreview() {
     KaigiTheme {
