+++ feature/sessions/src/main/java/io/github/droidkaigi/confsched2023/sessions/component/TimetableListItem.kt
@@ -1,14 +1,8 @@
 package io.github.droidkaigi.confsched2023.sessions.component
 
-import android.os.Build
-import androidx.compose.animation.graphics.ExperimentalAnimationGraphicsApi
-import androidx.compose.animation.graphics.res.animatedVectorResource
-import androidx.compose.animation.graphics.res.rememberAnimatedVectorPainter
-import androidx.compose.animation.graphics.vector.AnimatedImageVector
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ExperimentalLayoutApi
@@ -33,10 +27,6 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -52,7 +42,6 @@ import io.github.droidkaigi.confsched2023.designsystem.preview.MultiLanguagePrev
 import io.github.droidkaigi.confsched2023.designsystem.preview.MultiThemePreviews
 import io.github.droidkaigi.confsched2023.designsystem.theme.KaigiTheme
 import io.github.droidkaigi.confsched2023.designsystem.theme.md_theme_light_outline
-import io.github.droidkaigi.confsched2023.feature.sessions.R
 import io.github.droidkaigi.confsched2023.model.TimetableItem
 import io.github.droidkaigi.confsched2023.model.TimetableItem.Session
 import io.github.droidkaigi.confsched2023.model.fake
@@ -65,7 +54,7 @@ import java.lang.Integer.max
 const val TimetableListItemTestTag = "TimetableListItem"
 const val TimetableListItemBookmarkIconTestTag = "TimetableListItemBookmarkIconTestTag"
 
-@OptIn(ExperimentalLayoutApi::class, ExperimentalAnimationGraphicsApi::class)
+@OptIn(ExperimentalLayoutApi::class)
 @Composable
 fun TimetableListItem(
     timetableItem: TimetableItem,
@@ -80,9 +69,7 @@ fun TimetableListItem(
     ) {
         Row(verticalAlignment = Alignment.CenterVertically) {
             FlowRow(
-                modifier = Modifier
-                    .weight(1F)
-                    .padding(top = 4.dp),
+                modifier = Modifier.weight(1F).padding(top = 4.dp),
                 horizontalArrangement = Arrangement.spacedBy(4.dp),
                 verticalArrangement = Arrangement.spacedBy(8.dp),
             ) {
@@ -96,15 +83,19 @@ fun TimetableListItem(
                     checkedContentColor = MaterialTheme.colorScheme.onSurface,
                 ),
             ) {
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
-                    AnimatedBookmarkIcon(
-                        isBookmarked = isBookmarked,
-                    )
-                } else {
-                    BookmarkIcon(
-                        isBookmarked = isBookmarked,
-                    )
-                }
+                Icon(
+                    imageVector = if (isBookmarked) {
+                        Icons.Filled.Bookmark
+                    } else {
+                        Icons.Outlined.BookmarkBorder
+                    },
+                    contentDescription = if (isBookmarked) {
+                        SessionsStrings.RemoveFromFavorites.asString()
+                    } else {
+                        SessionsStrings.AddToFavorites.asString()
+                    },
+                    modifier = Modifier.padding(top = 4.dp),
+                )
             }
         }
         Spacer(modifier = Modifier.size(5.dp))
@@ -187,53 +178,6 @@ fun TimetableListItem(
         Spacer(modifier = Modifier.size(15.dp))
         Divider()
     }
-}
-
-@OptIn(ExperimentalAnimationGraphicsApi::class)
-@Composable
-fun AnimatedBookmarkIcon(
-    isBookmarked: Boolean,
-    modifier: Modifier = Modifier,
-) {
-    var atEnd by remember { mutableStateOf(false) }
-    val animatedBookmarkIcon = AnimatedImageVector.animatedVectorResource(
-        id = if (isBookmarked && atEnd) {
-            R.drawable.animated_bookmark_icon_reverse
-        } else {
-            R.drawable.animated_bookmark_icon
-        }
-    )
-    Icon(
-        painter = rememberAnimatedVectorPainter(animatedBookmarkIcon, atEnd),
-        contentDescription = if (isBookmarked) {
-            SessionsStrings.RemoveFromFavorites.asString()
-        } else {
-            SessionsStrings.AddToFavorites.asString()
-        },
-        modifier = modifier
-            .padding(top = 4.dp)
-            .clickable { atEnd = atEnd.not() },
-    )
-}
-
-@Composable
-fun BookmarkIcon(
-    isBookmarked: Boolean,
-    modifier: Modifier = Modifier,
-) {
-    Icon(
-        imageVector = if (isBookmarked) {
-            Icons.Filled.Bookmark
-        } else {
-            Icons.Outlined.BookmarkBorder
-        },
-        contentDescription = if (isBookmarked) {
-            SessionsStrings.RemoveFromFavorites.asString()
-        } else {
-            SessionsStrings.AddToFavorites.asString()
-        },
-        modifier = modifier.padding(top = 4.dp),
-    )
 }
 
 @MultiThemePreviews
